# caliborn.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "caliborn",
    "description": "API server for lumiRadio",
    "contact": {
      "name": "cozygalvinism"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/crud/_meta": {
      "get": {
        "tags": [
          "crate::routes::admin::crud"
        ],
        "summary": "Every resource name the registry serves.",
        "operationId": "meta",
        "responses": {
          "200": {
            "description": "Registered resource names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_admin_crud` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/admin/crud/{resource}": {
      "get": {
        "tags": [
          "crate::routes::admin::crud"
        ],
        "summary": "List a page of records for `resource`.",
        "operationId": "list",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "description": "Resource name, as listed by `GET /admin/crud/_meta`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of records; shape depends on the resource",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_admin_crud` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "crate::routes::admin::crud"
        ],
        "summary": "Create a record. The body is validated against the resource's create DTO.",
        "operationId": "create",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "description": "Resource name, as listed by `GET /admin/crud/_meta`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Resource-specific create payload; see `GET /admin/crud/{resource}/_schema`",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The created record",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Payload did not match the resource's create DTO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_admin_crud` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/admin/crud/{resource}/_schema": {
      "get": {
        "tags": [
          "crate::routes::admin::crud"
        ],
        "summary": "Column metadata for `resource` — drives generic admin form rendering.",
        "operationId": "schema",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "description": "Resource name, as listed by `GET /admin/crud/_meta`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One entry per column",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldMeta"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_admin_crud` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/admin/crud/{resource}/{id}": {
      "get": {
        "tags": [
          "crate::routes::admin::crud"
        ],
        "summary": "Read a single record by primary key.",
        "operationId": "read",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "description": "Resource name, as listed by `GET /admin/crud/_meta`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Primary key of the record, as a string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record, or `null` if no row has that id",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Id could not be parsed into the resource's key type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_admin_crud` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "crate::routes::admin::crud"
        ],
        "summary": "Partially update a record. Omitted fields keep their current value.",
        "operationId": "edit",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "description": "Resource name, as listed by `GET /admin/crud/_meta`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Primary key of the record, as a string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Resource-specific update payload; every field is optional",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated record",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Payload did not match the resource's update DTO, or the id was unparseable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_admin_crud` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "crate::routes::admin::crud"
        ],
        "summary": "Delete a record by primary key.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "description": "Resource name, as listed by `GET /admin/crud/_meta`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Primary key of the record, as a string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record deleted"
          },
          "400": {
            "description": "Id could not be parsed into the resource's key type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_admin_crud` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "An error response containing details about the problem",
        "required": [
          "message",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A machine-readable error code identifying the error type.\n\nCommon values include \"Bad Request\", \"Unauthorized\", \"Internal Server Error\",\nand \"Not Found\".",
            "examples": [
              "Bad Request",
              "Unauthorized",
              "Internal Server Error"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable message describing the error.",
            "examples": [
              "Invalid authorization code"
            ]
          }
        },
        "examples": [
          {
            "error": "Unauthorized",
            "message": "Invalid authorization code"
          }
        ]
      },
      "FieldMeta": {
        "type": "object",
        "description": "One column of an admin resource, as reported by\n`GET /admin/crud/{resource}/_schema`.",
        "required": [
          "name",
          "column_type",
          "nullable",
          "is_primary_key"
        ],
        "properties": {
          "column_type": {
            "type": "string"
          },
          "is_primary_key": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "nullable": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "user_api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      },
      "user_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}